legacy-python/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
fpdf==1.7.2
idna==3.11
psutil==7.2.2
requests==2.32.5
speedtest-cli==2.1.3
urllib3==2.6.3
python-dotenv==1.0.1